requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
psycopg2-binary==2.9.9
python-dotenv==1.0.0
openai==0.28.1
langchain==0.0.335
pydantic==2.4.2
requests==2.31.0
